test: add unit tests for workday document retrieval and validation
