+++ api/tests/workdays/documents/mod.rs
@@ -1,0 +1,3 @@
+pub mod get_workday_document_by_month;
+pub mod get_workday_document_years;
+pub mod get_workday_documents;
